Refactor Coordinate Class
